pyproject.toml
[tool.poetry]
name = "qualifyai"
version = "0.1.0"
description = ""
authors = ["Leandro S. Silva"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-api-python-client = "^2.149.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
